linters: add a second path in swift-format for Windows
This adds a second path of execution specifically for Windows.  The
second path is required due to changes in how the action can be used.
In particular:

- `prefix` cannot be supported as we do not go through a shell
- `args` is currently unsupported as the style must be different

In order to regain the ability add supplementary arguments with execute,
the `args` parameter must change from a string to an array of strings.